project/blockchain/temp/requirements.txt
azure-functions==1.23.0
iota-sdk==1.1.4
psycopg2-binary==2.9.10
